fse: reorganize block-inserter-modifications files
