--- apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/index.js
@@ -1,5 +1,0 @@
-/**
- * Internal dependencies
- */
-import './new-blocks-showcase';
-import './tips';
