+++ apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/tips.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { debounce, get, filter, deburr, lowerCase, includes, uniq } from 'lodash';
+import { debounce } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -9,40 +9,12 @@ import { debounce, get, filter, deburr, lowerCase, includes, uniq } from 'lodash
 import { useState } from '@wordpress/element';
 import { __experimentalInserterMenuExtension as InserterMenuExtension } from '@wordpress/block-editor';
 import { registerPlugin } from '@wordpress/plugins';
-import { Tip } from '@wordpress/components';
-
-import './index.scss';
 
 /**
  * Internal dependencies
  */
-import tips from './tips';
-
-function ContextualTip( { searchTerm, random = false } ) {
-	if ( ! searchTerm ) {
-		return null;
-	}
-
-	if ( ! tips.length ) {
-		return null;
-	}
-
-	const normalizedSearchTerm = deburr( lowerCase( searchTerm ) ).replace( /^\//, '' );
-
-	const foundTips = filter(
-		tips,
-		( { keywords, permission } ) =>
-			permission() &&
-			filter( uniq( keywords ), ( keyword ) => includes( normalizedSearchTerm, keyword ) ).length
-	);
-
-	if ( ! foundTips.length ) {
-		return null;
-	}
-
-	const index = random ? Math.floor( Math.random() * foundTips.length ) : 0;
-	return <Tip> { get( foundTips, [ index, 'description' ] ) }</Tip>;
-}
+import ContextualTip from './tips/contextual-tip';
+import './tips/style.scss';
 
 const ContextualTips = function () {
 	const [ debouncedFilterValue, setFilterValue ] = useState( '' );
@@ -68,7 +40,7 @@ const ContextualTips = function () {
 
 // Check if the experimental slot is available before to register plugin.
 if ( typeof InserterMenuExtension !== 'undefined' ) {
-	registerPlugin( 'insert-menu-contextual-tips', {
+	registerPlugin( 'block-inserter-contextual-tips', {
 		render() {
 			return <ContextualTips />;
 		},
