+++ apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/tips/contextual-tip.js
@@ -1,0 +1,40 @@
+/**
+ * External dependencies
+ */
+import { get, filter, deburr, lowerCase, includes, uniq } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { Tip } from '@wordpress/components';
+
+/**
+ * Internal dependencies
+ */
+import tipsList from './list';
+
+export default function ( { searchTerm, random = false } ) {
+	if ( ! searchTerm ) {
+		return null;
+	}
+
+	if ( ! tipsList.length ) {
+		return null;
+	}
+
+	const normalizedSearchTerm = deburr( lowerCase( searchTerm ) ).replace( /^\//, '' );
+
+	const foundTips = filter(
+		tipsList,
+		( { keywords, permission } ) =>
+			permission() &&
+			filter( uniq( keywords ), ( keyword ) => includes( normalizedSearchTerm, keyword ) ).length
+	);
+
+	if ( ! foundTips.length ) {
+		return null;
+	}
+
+	const index = random ? Math.floor( Math.random() * foundTips.length ) : 0;
+	return <Tip> { get( foundTips, [ index, 'description' ] ) }</Tip>;
+}
