+++ apps/full-site-editing/full-site-editing-plugin/block-inserter-modifications/tips/list.js
@@ -1,3 +1,4 @@
+
 /**
  * External dependencies
  */
